Refactor Future.hook
Now when a hooked Future is cancelled, its disposal Future is
still executed, and immediately cancelled as well. This creates
a much better chance for resources to be disposed, and allows
a developer to forcefully get rid of resources in a much more
logical place: The disposal cancellation function instead of
the consumpsion cancellation function.
Also improve toString and probably performance.