Change the way tests are executed
We no longer compile our unit-testing code. Instead, we bundle
the tests with Rollup, instrumenting them for coverage in the
process. We then run the tests natively on Node v7. We trust
that Buble will do its job and transpile the code such that if
the tests pass on the native code, they do on the transpiled.

This was the only approach in which I was able to get both:

* Stack traces reporting the correct origin of thrown errors
* Test coverage reports highlighting the correct ES6 code

Now that code coverage reports are finally intelligable, I can
move on to work on #88.