Remove type-checking from chainRec
Future.chainRec is no longer a "public" (documented)
function, as users are now able to achieve the same
using Future.chain. The only reason we keep it
around is for abstractions over Fantasy Land
ChainRec and Static Land ChainRec. Type checking
is merely a means to friendlier error messages and
friendly error messages are for users. Taking this
feature out increases the performance of code which
builds on ChainRec.